refactor: Move /me route to /users router
